Fix some more prototypes (size_t != necessarily unsigned int).
--HG--
extra : source : a0cb877563230b965060390110a14aea8a6091c5

+++ Library/libs/bcmp.c
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <sys/types.h>
 
-int bcmp(const void *dest, const void *src, int len)
+int bcmp(const void *dest, const void *src, size_t len)
 {
    return memcmp(dest, src, len);
 }
